feat(logentrysetup): Add checkboxes for visibility control
Adds checkboxes to the Log Entry Setup page to control the visibility of log entries in the admin and teacher portals, and on the teacher portal entry form.  Modified `plugin.xml` to update the version and `logentrysetup.enhanced_logs.content.footer.txt` to add the new form controls. Functionality to come in future commits.